refactor: class Context 的ioProcess()函数返回值从uint16_t改为int

+++ Coap/CoapPlusPlus/src/Context.cc
@@ -29,7 +29,7 @@ namespace CoapPlusPlus {
 //     }
 // }
 
-uint16_t Context::ioProcess(int waitMs)
+int Context::ioProcess(int waitMs)
 {
     if(isReady() == false) {
         throw DataNotReadyException("No endpoint or session added, no network I/O possible.");
